+++ tests/testthat/test-dataset-stream.R
@@ -11,9 +11,9 @@ mock_stream_rows <- data.frame(
 )
 
 with_mock_crunch({
-    ds <- loadDataset("test ds")   ## has 2 messages waiting, 4 rows received
+    ds <- loadDataset("streaming test ds")   ## has 2 messages waiting, 4 rows received
     ds2 <- loadDataset("an archived dataset", kind="archived") ## Has no streams mock
-    ds3 <- loadDataset("ECON.sav") ## has 0 messages waiting, 0 rows received
+    ds3 <- loadDataset("streaming no messages") ## has 0 messages waiting, 0 rows received
     test_that("pendingStream gets pending messages", {
         expect_equal(pendingStream(ds), 2)
         expect_GET(pendingStream(ds2), 'https://app.crunch.io/api/datasets/2/stream')
@@ -22,20 +22,20 @@ with_mock_crunch({
     test_that("streamRows streams rows", {
         expect_equal(streamRows(ds, data=data.frame()), ds)
         expect_POST(streamRows(ds, data=mock_stream_rows),
-                    'https://app.crunch.io/api/datasets/1/stream/',
+                    'https://app.crunch.io/api/datasets/1streaming/stream/',
                     '{"birthyr":0.5775,"gender":2,"subvar1":2,"subvar2":1,"subvar3":1,',
                     '"textVar":"a","starttime":"1955-12-28"}\n',
                     '{"birthyr":0.5775,"gender":1,"subvar1":2,"subvar2":1,"subvar3":1,',
                     '"textVar":"b","starttime":"1955-12-29"}')
         expect_POST(streamRows(ds[1,], data=mock_stream_rows),
-                    'https://app.crunch.io/api/datasets/1/stream/',
+                    'https://app.crunch.io/api/datasets/1streaming/stream/',
                     '{"birthyr":0.5775,"gender":2,"subvar1":2,"subvar2":1,"subvar3":1,',
                     '"textVar":"a","starttime":"1955-12-28"}')
     })
 
     test_that("appendStream", {
         expect_POST(appendStream(ds),
-                    'https://app.crunch.io/api/datasets/1/batches/',
+                    'https://app.crunch.io/api/datasets/1streaming/batches/',
                     '{"element":"shoji:entity","body":{',
                     '"type":"ldjson","stream":null}')
         expect_message(appendStream(ds3), "There's no pending stream data to be appended.")
@@ -52,7 +52,8 @@ stream_rows <- data.frame(
     v6 = c(1, 1)
 )
 with_test_authentication({
-    ds <- newDataset(df)
+    metadata <- fromJSON(file.path("dataset-fixtures", "streaming.json"), simplifyVector=FALSE)
+    ds <- suppressMessages(uploadMetadata(metadata))
     test_that("streamRows streams rows", {
         expect_equal(pendingStream(ds), 0)
         ds <- streamRows(ds, data=stream_rows)
@@ -61,7 +62,7 @@ with_test_authentication({
 
     test_that("appendStream appends all pending rows", {
         ds <- appendStream(ds)
-        expect_equal(nrow(ds), 22)
+        expect_equal(nrow(ds), 2)
         expect_equal(pendingStream(ds), 0)
     })
 })