+++ NEWS.md
@@ -1,5 +1,6 @@
 ### crunch 1.18.3 (under development)
 * Now Crunch authentication email and password can be stored in and read from the environmental variables `R_CRUNCH_EMAIL` and `R_CRUNCH_PW` respectively.
+* Fix `bases()` when called on a univariate statistic (#124)
 
 ### crunch 1.18.2
 * `makeWeight` allows you to generate new weighting variables based on categorical variables (#80).